backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
autoflake==2.1.1
autopep8==2.0.4
black==23.3.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
Django==3.2.14
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.0
djhtml==3.0.6
drf-yasg==1.21.7
Faker==12.0.1
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
mixer==7.2.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
promise==2.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==2.0.5
